schema/dlt/requirements-dbt.txt
dbt-duckdb>=1.9.0
